<Test> Update the unit tests of mixing.
